refactor(cosmic-swingset): introduce inboundQueue
