+++ packages/cosmic-swingset/src/chain-main.js
@@ -24,6 +24,7 @@ import { BridgeId as BRIDGE_ID } from '@agoric/internal';
 import stringify from './json-stable-stringify.js';
 import { launch } from './launch-chain.js';
 import { getTelemetryProviders } from './kernel-stats.js';
+import { makeQueue } from './make-queue.js';
 
 // eslint-disable-next-line no-unused-vars
 let whenHellFreezesOver = null;
@@ -112,81 +113,6 @@ const makeChainStorage = (call, prefix = '', options = {}) => {
     onAbort: resetCache,
   });
   return { ...buffered, commit, abort };
-};
-
-/**
- * Create a queue backed by chain storage.
- *
- * The queue uses the following storage layout, prefixed by `prefix`, such as
- * `actionQueue.`:
- * - `<prefix>head`: the index of the first entry of the queue.
- * - `<prefix>tail`: the index *past* the last entry in the queue.
- * - `<prefix><index>`: the contents of the queue at the given index.
- *
- * For the `actionQueue`, the Cosmos side of the queue will push into the queue,
- * updating `<prefix>tail` and `<prefix><index>`.  The JS side will shift the
- * queue, updating `<prefix>head` and reading and deleting `<prefix><index>`.
- *
- * Parallel access is not supported, only a single outstanding operation at a
- * time.
- *
- * @param {(obj: any) => any} call send a message to the chain's storage API and
- * receive a reply
- * @param {string} [prefix] string to prepend to the queue's storage keys
- */
-const makeChainQueue = (call, prefix = '') => {
-  const storage = makeChainStorage(call, prefix);
-  const queue = {
-    push: obj => {
-      const tail = storage.get('tail') || 0;
-      storage.set('tail', tail + 1);
-      storage.set(`${tail}`, obj);
-      storage.commit();
-    },
-    /** @type {Iterable<unknown>} */
-    consumeAll: () => ({
-      [Symbol.iterator]: () => {
-        let done = false;
-        let head = storage.get('head') || 0;
-        const tail = storage.get('tail') || 0;
-        return {
-          next: () => {
-            if (done) return { done };
-            if (head < tail) {
-              // Still within the queue.
-              const headKey = `${head}`;
-              const value = storage.get(headKey);
-              storage.delete(headKey);
-              head += 1;
-              return { value, done };
-            }
-            // Reached the end, so clean up our indices.
-            storage.delete('head');
-            storage.delete('tail');
-            storage.commit();
-            done = true;
-            return { done };
-          },
-          return: () => {
-            if (done) return { done };
-            // We're done consuming, so save our state.
-            storage.set('head', head);
-            storage.commit();
-            done = true;
-            return { done };
-          },
-          throw: err => {
-            if (done) return { done };
-            // Don't change our state.
-            storage.abort();
-            done = true;
-            throw err;
-          },
-        };
-      },
-    }),
-  };
-  return queue;
 };
 
 export default async function main(progname, args, { env, homedir, agcc }) {
@@ -323,9 +249,11 @@ export default async function main(progname, args, { env, homedir, agcc }) {
         toChainShape: exportMailbox,
       },
     );
-    const actionQueue = makeChainQueue(
-      msg => chainSend(portNums.storage, msg),
-      `${STORAGE_PATH.ACTION_QUEUE}.`,
+    const actionQueue = makeQueue(
+      makeChainStorage(
+        msg => chainSend(portNums.storage, msg),
+        `${STORAGE_PATH.ACTION_QUEUE}.`,
+      ),
     );
     function setActivityhash(activityhash) {
       const msg = stringify({
